StackFrame: remove off-by-one from `lookup`
We have long been doing an off-by-one computation in `lookup`.
This is often convenient, but a bit confusing in cases where you then have to then reverse the adjustment for the correction.
As we continue working towards more precise profile results, this becomes somewhat more significant.
So try to cleanup some accumulated code-cruft around it and in related code.